geoecu solver

+++ app/models/simulation.rb
@@ -8,10 +8,8 @@ class Simulation < ActiveRecord::Base
 
   def store_results(solver)
     return unless solver.finished? && solver.optimal?
-    FileUtils.mkdir_p(results_path)
-    EXT.each do |x|
-      FileUtils.cp(solver.file(x),File.join(results_path,"sim.#{x}"))
-    end
+    FileUtils.mkdir_p(results_path) unless File.exists?(results_path)
+    EXT.each { |x| FileUtils.cp(solver.file(x),file(x)) }
   end
 
   private
@@ -20,12 +18,13 @@ class Simulation < ActiveRecord::Base
     File.join(RAILS_ROOT,'public','files','sim',self.id.to_s)
   end
 
+  def file(ext)
+    File.join(results_path,"sim.#{ext}")
+  end
+
   def delete_results
-    EXT.each do |x|
-      f = File.join(results_path,"sim.#{x}")
-      File.delete(f) if File.exists?(f)
-    end
-    FileUtils.rmdir(results_path)
+    EXT.each {|x| File.delete(file(x)) if File.exists?(file(x)) }
+    FileUtils.rmdir(results_path)  if File.exists?(results_path)
   end
 
 end
